python/requirements.txt
earthengine_api==0.1.369
geopandas==0.13.2
GOSTnets==1.1.0
GOSTRocks==0.2.0
matplotlib==3.7.2
networkx==3.1
numpy==1.25.2
ogr==0.49.2
osmnx==1.9.2
osr==0.0.1
pandana==0.7
pandas==2.0.3
pyquadkey2==0.2.2
PyYAML==6.0.1
rasterio==1.3.8
Requests==2.32.2
scipy==1.11.2
Shapely==2.0.3
xarray==2023.8.0
